refactor(core): extract language workflow owner
