+++ crates/tokmd-core/src/workflows/mod.rs
@@ -1,0 +1,5 @@
+//! Public workflow facade owner modules.
+
+mod lang;
+
+pub use lang::{lang_workflow, lang_workflow_from_inputs};
